FEAT: 관심매물 등록/삭제/리스트 추가

+++ VabangSpringBoot/src/main/java/com/ssafy/controller/InterestSaleController.java
@@ -1,54 +1,64 @@
 package com.ssafy.controller;
 
-import com.ssafy.domain.InterestSaleInfo;
+import com.ssafy.domain.interest.InterestSaleDTO;
+import com.ssafy.domain.interest.InterestSaleInfo;
 import com.ssafy.service.InterestSaleService;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RequestMapping("/interest")
 @RestController
+@CrossOrigin("*")
 public class InterestSaleController {
 
     @Autowired
     InterestSaleService interestSaleService;
 
-    @GetMapping("/show/{userId}")
-    public List<InterestSaleInfo> showInterests(@PathVariable String userId) { //@SessionAttribute User user){
-        //if (user == null) return null;
-        //String userId = user.getId();
-        return interestSaleService.showInterests(userId);
-    }
-
     @PostMapping("/insert")
-    int insertInterest(@RequestParam("userId") String userId, @RequestParam("dealNo") String dealNo) { //@SessionAttribute User user, @RequestParam String dealNo){
-        //if (user == null) return 0;
-        //String userId = user.getId();
+    int insertInterest(@RequestBody String aptCode){//(@SessionAttribute User user, @RequestParam String aptCode){
+        /*
+        if (user == null) return 0;
+        String userId = user.getUserId();
+        */
+        String userId = "admin";
         Map<String, String> dealUserMap = new HashMap<>();
         dealUserMap.put("userId", userId);
-        dealUserMap.put("dealNo", dealNo);
+        dealUserMap.put("aptCode", aptCode);
         return interestSaleService.insertInterest(dealUserMap);
     }
 
-    @DeleteMapping("/delete")
-    int deleteInterest(@RequestParam("userId") String userId, @RequestParam("dealNo") String dealNo) { //@SessionAttribute User user, @RequestParam String dealNo){
-        //if (user == null) return 0;
-        //String userId = user.getId();
+    @DeleteMapping("/delete/{aptCode}")
+    int deleteInterest(@PathVariable String aptCode) {//(@SessionAttribute User user, @RequestParam String aptCode){
+        /*
+        if (user == null) return 0;
+        String userId = user.getUserId();
+         */
+        String userId = "admin";
         Map<String, String> dealUserMap = new HashMap<>();
         dealUserMap.put("userId", userId);
-        dealUserMap.put("dealNo", dealNo);
+        dealUserMap.put("aptCode", aptCode);
         return interestSaleService.deleteInterest(dealUserMap);
     }
 
+    @GetMapping("/list")
+    List<InterestSaleInfo> getUserInterestList() {
+        String userId = "admin";
+        if (userId == null) return null;
 
+        List<InterestSaleInfo> list = interestSaleService.getUserInterestList(userId);
+        System.out.println(list);
+        return list;
+    }
 
+    @GetMapping("/getInterests")
+    List<InterestSaleDTO> getInterests(){
+        String userId = "admin";
+        if (userId == null) return null;
 
+        List<InterestSaleDTO> list = interestSaleService.getInterests(userId);
+        return list;
+    }
 }
